Added log4j.properties, revised unit and int tests
